Bug solved in GenTable() function: mismatch between GO ids and term Description
The weight and elim algorithms have been optimize. The weight algorithm runs almost twice as fast now. 

New function added: nodesInInducedGraph() - similar with inducedGraph() but it just returns the nodes in the graph as a characher vector and not a graph object.

New feature added: When building the topGOdata object, the user can prune the GO graph such that the leafs in the graph have at least a number of annotated genes. By adding "nodeSize = k" in the constructor of the topGOdata object, only GO terms with more than k annotated genes will be used. By default, nodeSize = 0.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_2_3/madman/Rpacks/topGO@35382 bc3139a8-67e5-0310-9ffc-ced21a209358